+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Unify backing code when getting last result and streaming results (better RAM usage for long sequences).
 * Add function to clean temporary files.
 * Initial set of unit tests.
+* Use bracket and cleanupProcess to dispose of resources.
 
 ## 0.1.2.1 -- 2020-11-08